requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
gradio==4.44.1
pydantic==2.7.1
pydantic-settings==2.2.1
supabase>=2.12.0
telethon==1.36.0
instaloader==4.13.1
transformers==4.41.0
torch>=2.3.0
torchvision>=0.18.0
ultralytics==8.2.0
pytesseract==0.3.10
Pillow==10.3.0
exifread==3.0.0
geopy==2.4.1
phonenumbers==8.13.37
email-validator==2.1.1
python-whois==0.9.4
reportlab==4.2.0
apscheduler==3.10.4
httpx
python-dotenv==1.0.1
python-multipart==0.0.9
websockets
